Convert forceLightningAppCreate
